requirements.txt
streamlit==1.22.0
pandas==1.5.3
plotly==5.14.1
requests==2.28.2
streamlit-extras==0.3.2
numpy==1.24.3
woocommerce==3.0.0
woocommerce
mailwizz-python-sdk
pandas
plotly
streamlit
